Use byte instead of int for consts. Maybe we should use just plain #define?

+++ NettigoKeypad.h
@@ -7,13 +7,12 @@
 class NG_Keypad
 {
   public:
-  static const int NONE    = 0;
-  static const int SELECT  = 1;
-  static const int LEFT    = 2;
-  static const int DOWN    = 3;
-  static const int UP      = 4;
-  static const int RIGHT   = 5;
-
+  static const byte NONE    = 0;
+  static const byte SELECT  = 1;
+  static const byte LEFT    = 2;
+  static const byte DOWN    = 3;
+  static const byte UP      = 4;
+  static const byte RIGHT   = 5;
   
 
   NG_Keypad(void);
